Move signing / verifying claims to the generic jwtsecrets package
